iniciar-analise-http/requirements.txt
functions-framework
flask
flask-cors
google-cloud-storage
google-cloud-firestore
pytz
